+++ src/tools/Built-in/Ahk2exe.data.ts
@@ -8,20 +8,6 @@ type TAhk2exeDataElement = {
     exp: readonly string[],
 };
 
-// TODO /*@Ahk2Exe-Keep\n$0\n*/
-// const _keep = {
-//     keyRawName: 'Keep',
-//     link: 'https://www.autohotkey.com/docs/v1/misc/Ahk2ExeDirectives.htm#IgnoreKeep',
-//     doc: 'The reverse is also possible, i.e. marking a code section to only be executed in the compiled script:',
-//     body: '/*@Ahk2Exe-Keep\n$0\n*/',
-//     exp: [
-//         '/*@Ahk2Exe-Keep',
-//         'MsgBox This message appears only in the compiled script',
-//         '*/',
-//         'MsgBox This message appears in both the compiled and unCompiled script',
-//     ],
-// };
-
 /**
  * after initialization clear
  */