+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Next 0.0.29(2023-03-XX)
 
 - fix: not format `/*@Ahk2Exe-Keep` block
+- feat: add hover at `/*@Ahk2Exe-Keep`
 - feat: add semantic-highlight of `MyString = This is a literal string.` equal sign operator (=)
 
   ```ahk